Restructured Code
Restructured the code. Created new Packages and moved classes around.
This is a working version of the application.

Also added new classes that will be used for the UI. UserInput will be
a class that deals with all user input and creates the different menus
and sub-menus that call functions from the other controller classes.