Refactored Code Structure
Refactored the code structure to meet the professors requirements.
Created classes for the presentation layer the service layer and the
DAO. The presentation layer will be the part of the application that
takes user input. That user input is passed to the Service Layer that
will manipulate Model objects. After the Service Layer is finished with
an object it will pass that object to the DAO which will have Database
CRUD operations.