Implemented CLI
Began work on the CLI. At this point the application is working
somewhat. Currently there is a bug when creating a new appointment for
a patient that already has an appointment with the same doctor. With
this bug duplicate entries are added to the Patient_Doctor JoinTable.
So for every appointment you create 2 additional duplicate entries are
added.

One big feature that still needs to be implemented is the deleting of
appointments. Not exactly sure how this will work just yet. Still
thinking.

Also updated the hibernate.xml file to drop and recreate the database
with each run. Note that the very first time the application is
executed you will first have to manually create the database.