Major Updates
Have most of the CLI. Still cannot delete an entity. Also need to find
a way to accept more than a single word for Rx description in Doctor
menu. Some of the last things that need to be worked on are the
toString methods just to make sure that everything looks good. Also
need to add some error checks when searching for Entities.